requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0